requirements.txt
-e .
pytest==5.3.5
numpy==1.18.1
scipy==1.4.1
pandas==1.0.1
matplotlib==3.1.3
seaborn==0.10.0
torchsummary==1.5.1
python-telegram-bot==12.4.2
Pillow==6.1.0
pandas==1.0.1
SimpleITK==1.2.4